added a new problem called ultra fast mathematician

+++ CodeForces/Calculating Function.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-    int n, sum=0;
+    long long n, sum=0;
     cin>>n;
 
     if (n%2==0)
